% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_tdl.R
\name{theme_tdl}
\alias{theme_tdl}
\title{A TDL Theme}
\usage{
theme_tdl()
}
\arguments{
\item{nothing}{}
}
\description{
This theme is for Twitter
}
\examples{
use theme() afterwards
}
\keyword{tdl}
\keyword{twitter,}
