% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_tdl.R
\name{theme_tdl_pitch}
\alias{theme_tdl_pitch}
\title{A Twitter Theme}
\usage{
theme_tdl_pitch()
}
\arguments{
\item{love}{}
}
\description{
This theme is for Twitter
}
\examples{
use theme() afterwards
}
\keyword{twitter}
