% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_twiiter.R
\name{theme_twitter}
\alias{theme_twitter}
\title{A Twitter Theme}
\usage{
theme_twitter()
}
\arguments{
\item{love}{}
}
\description{
This theme is for Twitter
}
\examples{
use theme() afterwards
}
\keyword{twitter}
