% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_twitter_pitch.R
\name{theme_twitter_pitch}
\alias{theme_twitter_pitch}
\title{A Twitter Theme}
\usage{
theme_twitter_pitch()
}
\arguments{
\item{love}{}
}
\description{
This theme is for Twitter
}
\examples{
use theme() afterwards
}
\keyword{twitter}
